backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
qdrant-client
cohere
tiktoken
